+++ CHANGELOG.md
@@ -40,7 +40,10 @@ changes.
   - Overall this results in transactions still to be submitted once per client,
     but requires signifanctly less book-keeping on the client-side.
 
-- Add **Blockfrost Mode** to `hydra-chain-observer`, to follow the chain via Blockfrost API.
+- Add blockfrost support to `hydra-chain-observer`, to follow the chain via Blockfrost API.
+
+- Fix `bench-e2e single` benchmarks and only use `--output-directory` to keep
+  the whole benchmark state.
 
 - Add `inlineDatumRaw` to transaction outputs on the `hydra-node` API.